+++ plugins/system/cwgears/fields/alert.php
@@ -32,10 +32,9 @@ class CWElementAlert extends CWElement
      * @param $name
      * @param $value
      * @param $node
-     * @param $control_name
      * @return null
      */
-    public function fetchElement($name, $value, &$node, $control_name)
+    public function fetchElement($name, $value, &$node)
     {
 
         return NULL;
@@ -45,11 +44,10 @@ class CWElementAlert extends CWElement
      * @param $label
      * @param $description
      * @param $node
-     * @param $control_name
      * @param $name
      * @return string
      */
-    public function fetchTooltip($label, $description, &$node, $control_name, $name)
+    public function fetchTooltip($label, $description, &$node, $name)
     {
         return '<div class="cw-message-block">'
             . '<div class="cw-message">'