+++ plugins/system/cwgears/fields/info.php
@@ -31,10 +31,9 @@ class CWElementInfo extends CWElement {
      * @param $name
      * @param $value
      * @param $node
-     * @param $control_name
      * @return null
      */
-    public function fetchElement($name, $value, &$node, $control_name) {
+    public function fetchElement($name, $value, &$node) {
 
         return NULL;
     }
@@ -43,11 +42,10 @@ class CWElementInfo extends CWElement {
      * @param $label
      * @param $description
      * @param $node
-     * @param $control_name
      * @param $name
      * @return string
      */
-    public function fetchTooltip($label, $description, &$node, $control_name, $name) {
+    public function fetchTooltip($label, $description, &$node, $name) {
         return '<div class="cw-message-block">'
                 . '<div class="cw-message">'
                 . '<p class="info">' . JText::_($label) . '</p>'