+++ plugins/system/cwgears/fields/inherit.php
@@ -32,10 +32,9 @@ class CWElementInherit extends CWElement {
      * @param $name
      * @param $value
      * @param $node
-     * @param $control_name
      * @return null
      */
-    public function fetchElement($name, $value, &$node, $control_name) {
+    public function fetchElement($name, $value, &$node) {
         return NULL;
     }
 
@@ -43,11 +42,10 @@ class CWElementInherit extends CWElement {
      * @param $label
      * @param $description
      * @param $node
-     * @param $control_name
      * @param $name
      * @return string
      */
-    public function fetchTooltip($label, $description, &$node, $control_name, $name) {
+    public function fetchTooltip($label, $description, &$node, $name) {
         
                 // Load version.php
         jimport('joomla.filesystem.file');