+++ plugins/system/cwgears/fields/upgrade.php
@@ -32,10 +32,9 @@ class CWElementUpgrade extends CWElement {
      * @param $name
      * @param $value
      * @param $node
-     * @param $control_name
      * @return null
      */
-    public function fetchElement($name, $value, &$node, $control_name) {
+    public function fetchElement($name, $value, &$node) {
         return NULL;
 
     }
@@ -44,11 +43,10 @@ class CWElementUpgrade extends CWElement {
      * @param $label
      * @param $description
      * @param $node
-     * @param $control_name
      * @param $name
      * @return string
      */
-    public function fetchTooltip($label, $description, &$node, $control_name, $name) {
+    public function fetchTooltip($label, $description, &$node, $name) {
 
                 return '<div class="cw-message-block">'
                 . '<div class="cw-message">'