+++ plugins/system/cwgears/libraries/CoalaWeb/src/Messages.php
@@ -12,7 +12,6 @@
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
- *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
@@ -21,9 +20,9 @@
 
 namespace CoalaWeb;
 
-use Joomla\CMS\Language\Text as JText;
-
 defined('_JEXEC') or die;
+
+use Joomla\CMS\Language\Text as JText;
 
 /**
  * Class Xml