+++ plugins/system/cwgears/libraries/CoalaWeb/src/StringHelper.php
@@ -12,7 +12,6 @@
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
- *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
@@ -30,8 +29,7 @@ use Normalizer;
  * Class StringHelper
  * @package CoalaWeb
  */
-class StringHelper
-	extends \Joomla\String\StringHelper
+class StringHelper extends \Joomla\String\StringHelper
 {
 	/**
 	 * Decode html entities in string or array of strings