+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 ## Change Log
 
+### Version 0.6.1 October 2020
+- **(A)** New modal article picker form field
+- **(A)** Three new classes for the CoalaWeb library
+- **(A)** New database table to help with ciphers
+- **(C)** Code updates and improvements to support PHP 7.4
+- **(C)** Code updates and improvements for CSS loading fields
+- **(C)** Code updates and improvements for base, alert and upgrade fields
+- **(C)** Code updates and improvements
+- **(L)** New and updated language strings
+
 ### Version 0.6.0 April 2020
 - **(C)** Improved text cleaning to remove page builder tags
 - **(C)** Code updates and improvements